requirements.txt
flask
mysql-connector-python
gunicorn
requests
pytest
